"ads2_m6_a1_1: 3/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS2-Assignments/module6/ADS-2Week-1/Solution.java
@@ -52,7 +52,7 @@ class PageRank {
                         // k++;
                     }
                     else {
-                        prvalue = (double)1 / dg.vertex();
+                        prvalue = 1.0 / dg.vertex();
                         lst.put(j, prvalue);  
                     }
                 }
